feat: extract game setup helpers
